sting: complete shadow resolution
The only pending cases are when the last component of the path is
a mountpoint, and an oops on umounting a tmpfs during shutdown.

When the path is a mountpoint, the permission check has to be
done on the parent dentry outside the mountpoint, and the child dentry
which is the mountpoint. However, the reference to the resolved
root dentry has to be dropped too.