sting: present different unionfs branches depending on resolution context
There are three different resolution contexts for the branches
adversarial (id 0) and non-adversarial (id 1).
1. ADV_RES - only adversarial branch should be shown.
2. ADV_NORMAL_RES - adversarial branch shown if available, else normal
					branch shown.
3. NORMAL_RES - only non-adversarial branch is shown.

Which of these three should be shown is set by shadow resolution based
on looking up adversary accessibility to the resolution that is about
to be done. Shadow resolution itself uses ADV_NORMAL_RES, rollback
uses ADV_RES, and launch does it out-of-band (technically, launch
should use ADV_RES also, but the full directory structure may not be
available in the adversarial branch and we may hit a dead-end, whereas
we really want to do a unionfs "copyup").

unionfs agnostic except create and delete.

General strategy to show different branches:
In d_revalidate(), recreate the dentry if its existing context is not
appropriate for our particular resolution context. To avoid
revalidating in the most common case (i.e.,) when we have resolved for
ADV_NORMAL_RES (as shadow resolution does), and want to use NORMAL_RES
when there is no adversarial object, we store an additional "tried"
branch index range (tbstart, tbend) in struct unionfs_dentry_info.

Changes to unionfs to support sting:
1. Disable copyup, whiteout, opaque directories.
2. Change lookup and revalidate to reconstruct dentries when invalid
	for the given resolution context (explained above)
3. If a victim deletes or renames adversarial branch object, delete or
	rename lower non-adversarial branch also (because there is no
	opacity or whiteout concept). This requires changes to
	unionfs_(rmdir, unlink, rename).
4. A victim process always creates on the non-adversarial branch.
	Victim process creation on the adversarial branch is always a
	vulnerability. This requires changes to unionfs_(create, link,
	symlink, mkdir, mknod).

Limitations:
1. A non-adversarial resource may not always be available to switch to
	once the vulnerability is marked and the adversarial resource
	deleted. In such a case, we fail. While
	this does not violate unix semantics (we only carry out legitimate
	adversarial actions), it may lead to unexpected program responses.